Issue #1362 : use np.intp types for indexing instead of np.int64

- According to numpy/numpy#4384 (comment), `np.intp` is the recommended dtype for indexing:
   "... use np.intp as dtype whenever things have to do with indexing or are logically related to 
   indexing/array sizes. This is the natural dtype for it and it will normally also be the fastest one.
- see https://docs.scipy.org/doc/numpy/user/basics.types.html
- see also PR #1391 for discussion
- This change should go some way towards increasing compatibility with i386/i696 (32 bit) platforms.

NOTE: No dedicated testing on 32 bit yet (see WIP PR #1392 )

+++ package/MDAnalysis/core/groups.py
@@ -419,7 +419,7 @@ class GroupBase(_MutableBase):
             ix, u = args
 
         # indices for the objects I hold
-        self._ix = np.asarray(ix, dtype=np.int64)
+        self._ix = np.asarray(ix, dtype=np.intp)
         self._u = u
         self._cache = dict()
 
@@ -2464,7 +2464,7 @@ class ComponentBase(_MutableBase):
         --------
         ix
         """
-        return np.array([self.ix])
+        return np.array([self.ix], dtype=np.intp)
 
 
 class Atom(ComponentBase):
@@ -2736,7 +2736,7 @@ class UpdatingAtomGroup(AtomGroup):
             ix = sum([sel.apply(bg) for sel in sels[1:]],
                      sels[0].apply(bg)).ix
         else:
-            ix = np.array([], dtype=np.int)
+            ix = np.array([], dtype=np.intp)
         # Run back through AtomGroup init with this information to remake ourselves
         super(UpdatingAtomGroup, self).__init__(ix, self.universe)
         self.is_uptodate = True
